% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{clue_gmt_from_list}
\alias{clue_gmt_from_list}
\title{Prepare gmt files from a list of genes}
\usage{
clue_gmt_from_list(up, down, name, drop_invalid = FALSE)
}
\description{
Prepare gmt files from a list of genes
}
