% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{clue_parse_result}
\alias{clue_parse_result}
\title{Parse result folder}
\usage{
clue_parse_result(
  path,
  score_level = c("cell", "summary"),
  score_type = c("ns", "tau"),
  result_type = c("pert", "pcl")
)
}
\arguments{
\item{path}{Path to Clue result gz file or extracted folder.}

\item{score_level}{Return result on cell level (cell) or summarized (summary)}

\item{score_type}{Return either normalized scores (ns) or CMap connectivity
scores (tau). See details.}

\item{result_type}{Return either perturbagen level information (pert) or
aggregated perturbagen class level information (pcl).}
}
\value{
A tibble in tidy format, with the score for a single
pertubation/gene set combination per row.
}
\description{
Clue calculates different scores.
\href{https://clue.io/connectopedia/connectivity_scores}{Connectopedia} gives
an overview of their interpretation.
}
