% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_api.R
\name{clue_prepare_deseq2}
\alias{clue_prepare_deseq2}
\title{Prepare a DESeq2 result for Clue}
\usage{
clue_prepare_deseq2(result_df, gene_set_name, alpha = 0.05)
}
\description{
Prepare a DESeq2 result for Clue
}
