% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_api.R
\name{clue_query_download}
\alias{clue_query_download}
\title{Download Clue job results}
\usage{
clue_query_download(clue_query, destination = NULL, api_key = NULL)
}
\arguments{
\item{clue_query}{Job ID or job parameters returned by
\code{\link{clue_query_submit}}}

\item{destination}{Path to download destination.}

\item{api_key}{Clue API key. Leave empty if it is saved in \code{~/.Renviron}}
}
\value{
Path to result tarball.
}
\description{
Given a job ID this function will download the results of a Clue job
as compressed tarball to the given location or to a temprary folder
}
