% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_api.R
\name{clue_query_wait}
\alias{clue_query_wait}
\title{Wait for query completion}
\usage{
clue_query_wait(clue_query, interval = 60, timeout = 3600,
  quiet = FALSE, api_key = NULL)
}
\description{
Wait for query completion
}
