% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_space}
\alias{gene_space}
\title{Gene space of CMap data}
\format{
A data frame with 20864 rows and 4 variables:
\describe{
\item{entrez_id}{Entrez gene ID}
\item{symbo}{HUGO gene symbol}
\item{type}{Gene detected directly (landmark) or inferred with different level of confidence}
\item{bing}{Indicator if gene is part of the BING space of best inferred genes}
}
}
\source{
\url{https://clue.io/command?q=/gene-space}
}
\usage{
gene_space
}
\description{
Gene space of CMap data
}
\keyword{datasets}
