% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_api.R
\name{list_jobs}
\alias{list_jobs}
\title{List jobs}
\usage{
list_jobs(api_key = NULL)
}
\arguments{
\item{api_key}{Clue API key. Leave empty if it is saved in \code{~/.Renviron}}
}
\value{
Data frame with job info from Clue.
}
\description{
List all non-deleted jobs submitted by this user.
}
