% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue_api.R
\name{clue_query_poll}
\alias{clue_query_poll}
\alias{clue_query_status}
\alias{clue_query_wait}
\title{Poll query job status}
\usage{
clue_query_poll(clue_query, api_key = NULL)

clue_query_status(clue_query, api_key = NULL, from_poll = FALSE)

clue_query_wait(
  clue_query,
  interval = 60,
  timeout = 3600,
  quiet = FALSE,
  api_key = NULL
)
}
\arguments{
\item{clue_query}{Job ID or job parameters returned by
\code{\link{clue_query_submit}}}

\item{api_key}{Clue API key. Leave empty if it is saved in \code{~/.Renviron}}

\item{from_poll}{If TRUE, \code{clue_query} is assumed to be output from
\code{clue_query_poll()} output. Otherwise it is assumed to be job ID.}

\item{interval}{Check every x seconds.}

\item{timeout}{Abort waiting after x seconds.}

\item{quiet}{Don't output periodic status updates.}
}
\value{
List of status codes from Clue.
}
\description{
Find out status of a job or wait for its completion.
}
\section{Functions}{
\itemize{
\item \code{clue_query_status}: Get job status

\item \code{clue_query_wait}: Wait for query completion
}}

